test(landing): mark and move the #2782 sync strict-JSON reproduction into tests/regression
tests/sync/test_strict_json_stdout.py::test_mission_create_json_strict_when_sync_skips_ingress
already carried @pytest.mark.regression (with a code-comment pin to #2782)
alongside several unrelated, passing strict-JSON-contract tests. Extract it
to tests/regression/test_issue_2782_sync_strict_json_ingress_skip.py, which
imports the source file's isolation harness (_worktree_src,
_build_isolated_home, _resolve_in_tree_specify_cli, _run_cli_isolated,
_stop_isolated_sync_daemon, _scaffold_minimal_kittify_repo) read-only rather
than duplicating it. Root-cause docstring and assertions preserved
verbatim; the test still reproduces its #2782 defect (RED) when collected
across the whole tree the way the real regression job runs it.
